requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
faiss-cpu
sentence-transformers
pypdf
numpy
python-multipart

rank_bm25
